--- touchvie_front/src/main/java/com/touchvie/touchvie_front/ui/activities/Carousel.java
@@ -1,16 +1,0 @@
-package com.touchvie.touchvie_front.ui.activities;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import com.touchvie.touchvie_front.R;
-
-public class Carousel extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_carousel);
-
-    }
-}
