+++ touchvie_front/src/main/java/com/touchvie/touchvie_front/ui/modules/TitleModule.java
@@ -4,7 +4,7 @@ import android.view.View;
 import android.widget.TextView;
 
 import com.touchvie.touchvie_front.R;
-import com.touchvie.touchvie_front.ui.Module;
+import com.touchvie.touchvie_front.ui.views.Module;
 
 /**
  * Created by rodrigo.moral on 16/09/2016.
@@ -21,15 +21,15 @@ public class TitleModule extends Module {
         mTitle = (TextView) v.findViewById(R.id.module_title_txt);
     }
 
-    public TextView getmTitle() {
+    public TextView getTitle() {
         return mTitle;
     }
 
-    public void setmTitle(TextView mTitle) {
+    public void setTitle(TextView mTitle) {
         this.mTitle = mTitle;
     }
 
     public void configure(TitleModule titleVH, int position) {
-        titleVH.getmTitle().setText("TITLE TEST");
+        titleVH.getTitle().setText("TITLE TEST");
     }
 }
